Preprocessed files loading
Added a dump method to save all the metadata of a patch. Optionally, if
"true" is passed as a second argument, it can also save the image with
the patch-lines removed. dump takes a string as it first argument which
is used as the file name to save the data.

patchy.load now takes a second argument that is metadata, if metadata is
specified then it will ignore the lines on the image and skip the
processing. patchy.load will still cut the lines from the image.

Added patchy.import, a function that does the same that patchy.load but
the image that passed as first argument mustn't have the patch lines.
This makes the second argument, metadata, mandatory.